Solution to Input and Print a Double

+++ Input and Print a Double@@6788d5012db25efec946f942/code.c
@@ -5,8 +5,8 @@ char* welcome() {
 }
 
 int main() {
-    double x;
-    scanf("%lf", &x);
-    printf("You entered: %lf", x);
+    float x;
+    scanf("%.4f", &x);
+    printf("You entered: %.4f", x);
     return 0;
 }
